RangeMoveStrafe Fix + Test Cleanup
Fixed RangeMoveStrafe in that it wouldn't react correctly on the left side.
Removed duplicate test codes we had as / moved test codes to the test directory.
