+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AutoBlueRight.java
@@ -116,7 +116,7 @@ public class AutoBlueRight extends MyOpMode {
             sleep(1000);
 
 
-        vfMovePerp('b', rangeL);
+        vfMovePerp('b', rangeL,1);
 
 
         rangeMovePID(6.25, rangeF);
