+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AutoRedDriveLeft.java
@@ -133,7 +133,7 @@ public class AutoRedDriveLeft extends MyOpMode {
 //        }
 //        sleep(1000);
 
-        rangeMoveStrafe(26.25, rangeR);
+        rangeMoveStrafe(26.25, rangeR, 0);
 
 //[
 //        try {
