+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Test/AutoTest.java
@@ -27,7 +27,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.firstinspires.ftc.teamcode.Autonomous;
+package org.firstinspires.ftc.teamcode.Test;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.util.ElapsedTime;
@@ -37,16 +37,16 @@ import org.firstinspires.ftc.robotcore.external.navigation.Velocity;
 import org.firstinspires.ftc.teamcode.Library.MyOpMode;
 
 /**
- * {@link rangeTest} gives a short demo on how to use the BNO055 Inertial Motion Unit (IMU) from AdaFruit.
+ * {@link AutoTest} gives a short demo on how to use the BNO055 Inertial Motion Unit (IMU) from AdaFruit.
  *
  * Use Android Studio to Copy this Class, and Paste it into your team's code folder with a new name.
  * Remove or comment out the @Disabled line to add this opmode to the Driver Station OpMode list
  *
  * @see <a href="http://www.adafruit.com/products/2472">Adafruit IMU</a>
  */
-@Autonomous(name = "TelemTest", group = "Sensor")
+@Autonomous(name = "Test File", group = "Sensor")
                             // Comment this out to add to the opmode list
-public class rangeTest extends MyOpMode {
+public class AutoTest extends MyOpMode {
     private ElapsedTime runtime = new ElapsedTime();
 
     @Override
@@ -57,20 +57,26 @@ public class rangeTest extends MyOpMode {
         // Set up our telemetry dashboard
         composeTelemetry();
         // Wait until we're told to go
+
         waitForStart();
         runtime.reset();
 /**---------------------------------------------------------------------------------------------------------------*/
         // Start the logging of measured acceleration
         imu.startAccelerationIntegration(new Position(), new Velocity(), 1000);
 
-        //Need to put the Auto together - Message Caleb for the details on how.
-
-        // Loop and update the dashboard
-        while (opModeIsActive()) {
-
-            telemetry.update();
-//        }
+        try {
+            turnPID(-80);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
         }
+        sleep(2000);
+
+        //Finish optimizing this Auto, then invert for the blue side.
+        // Loop and update the dashboard
+//        while (opModeIsActive()) {
+//
+//            telemetry.update();
+//        }
 
     }
 }
